feat: Implement user reviews and enhance UI/UX
This commit introduces a user review system, allowing users to view ratings and reviews for fishers and buyers. It also includes significant UI/UX enhancements, such as changing price fields to integers, updating app branding, and improving navigation feedback.

Key changes:
-   **User Reviews & Ratings:**
    -   **New Screens (`fisher_review_screen.dart`, `buyer_review_screen.dart`):** Created dedicated screens to display a user's average rating, a rating distribution summary, and a list of individual reviews.
    -   **Navigation:** Users can now tap on a seller/buyer card in the `OfferDetails` screen to navigate to their review page.
    -   **Animated Ratings:** Replaced static star icons with the `animated_rating_stars` package for a more dynamic and interactive rating display in `RatingCard`, `ReviewCard`, and the rating modal.
    -   **Routing:** Added new routes (`/reviews/:userId`) for both buyer and fisher roles to handle navigation to review screens.

-   **Data & State Management:**
    -   **Integer Pricing:** Refactored data models (`Catch`, `Offer`), repositories, BLoCs, and UI components to use `int` for all price-related fields (`price`, `pricePerKg`, `total`) instead of `double`. This simplifies input and display.
    -   **Number Input Fields:** Updated `NumberInputField` and dialogs to enforce integer-only input for prices, preventing decimal entries.
    -   **Notification Badges:**
        -   Added notification counts to the "Offers" and "Messages" tabs on the `NotificationsScreen` to indicate pending updates.
        -   Implemented a badge on the `CatchDetails` screen's "Offers" tab to show the number of offers needing attention.

-   **UI/UX Enhancements:**
    -   **App Branding:**
        -   Updated the Android application label to "Siren Marketplace".
        -   Replaced all old `ic_launcher` (Android) and `AppIcon` (iOS) assets with new, branded application icons.
    -   **Page Titles:** Created a reusable `PageTitle` widget and applied it across multiple screens (`OfferDetails`, `CatchDetails`, `NotificationsScreen`, `AccountInfo`) for consistent styling.
    -   **Clickable User Cards:** The seller/buyer profile section in `OfferDetails` is now a clickable card that navigates to their reviews.
    -   **Catch Details Display:** Improved the `CatchDetails` and `ForSaleCard` to correctly display "Size" for prawns and "Average Size" in cm for other species.

-   **Dependencies:**
    -   Added the `animated_rating_stars: ^1.0.1` package for improved rating visualization.

+++ lib/core/models/catch.dart
@@ -13,8 +13,8 @@ class Catch extends Equatable {
   final String datePosted;
   final double initialWeight;
   final double availableWeight;
-  final double pricePerKg;
-  final double total;
+  final int pricePerKg;
+  final int total;
   final String size;
   final String market;
   final List<String> images;
@@ -81,8 +81,8 @@ class Catch extends Equatable {
     double? availableWeight,
     CatchStatus? status,
     List<Offer>? offers,
-    double? pricePerKg,
-    double? total,
+    int? pricePerKg,
+    int? total,
   }) {
     return Catch(
       id: id,
@@ -128,8 +128,8 @@ class Catch extends Equatable {
     datePosted: m['date_created'] as String,
     initialWeight: (m['initial_weight'] as num).toDouble(),
     availableWeight: (m['available_weight'] as num).toDouble(),
-    pricePerKg: (m['price_per_kg'] as num).toDouble(),
-    total: (m['total'] as num).toDouble(),
+    pricePerKg: (m['price_per_kg'] as num).toInt(),
+    total: (m['total'] as num).toInt(),
     size: m['size'] as String,
     market: m['market'] as String,
     species: Species(
